Add first version of the matrix multiplication lab
The tasks are currently computing groups of rows from the final result of the matrix multiplication.
If the rows of the result matrix cannot be exactly divided by the maximum possible number of threads supported by the system, the last thread computes the remaining number of rows (numberOfRowsComputedByOneThread + rows modulo numberOfRowsComputedByOneThread).